This commit fixes blackberry/BB10-WebWorks-Framework#123 that was introduced when we added  plugins. By doing so we changed the URI addresses used for our APIs. To address this, the bridge plugin was changed to a default plugin.
This changes the plugin loading logic, it now will use the default
plugin when the one given is undefined instead of throwing a 404.

This also includes changes so that the default plugin will throw a 404
if something doesn't exist and only after this check for Whitelisting.

Fixes code to not work based on exception