Adds preprocessing hooks into the build system.
The syntax for preprocessing is -
//@if <defines>
  ...
//@else
  ...
//@endif

<defines> are comma separated values such as test,qnxcar
The else block is optional.

Preprocessing the content means neither jake test or jake lint can pass
before the content is preprocessed. To solve this issue, during lint or
test a temp folder called 'temp' is created in the root and the appropriate
folders are copied there. After this jake lint is run from the temp folder
and jake test is run from the root but the specs are limited to the
specs in the temp/test folder. This completely solves the pathing issue
and any errors in jake lint or jake test are reported relatively and the
user does not need to be aware of the existence of the temp folder.

Also uncovers a jasmine bug that has not yet been merged into jasmine.
Basically an if check is needed to avoid throwing an exception during
running the tests.

The details of the bug can be found here, although the complete reason
for its occurence is unknown.
https://github.com/mhevery/jasmine-node/pull/5

This has been worked around by simply modifying a prototype of
jasmine-node to add the check. Details can be found in
dependencies/jasmine-node/jasmine-node.js