This introduces new wrappers in automation extension to allow taking a screenshot and comparing screenshots.
Developers taking screenshots can accomplish it using either the API
providing a path or using the "Take Screen Shot" button provided on the
whitel overlay shown by using "showOverlay()" on the automated spec
runner page.

If the button is used then developers can grab this newly taken screen
shot using new jake task. jake grab-image[ip,file-name] where file-name
is the name you want to name the file in the repo. The file is then
copied to test/data/automation/ReferenceImages.

For comparing images only the name of the image needs to passed in and
the current screenshot is compared against an image with that file name
on device under /accounts/1000/shared/misc/PuppetMaster/ReferenceImages.

To make this process more convinent for developers it augments new agents to the jake deploy-automation.
Namely automation-interface and PuppetMasterAgent have been added to the
jake task. It also copies all reference images from the repo to device.

Take away -
* use jake deploy-automation
* wait a few minutes and deploy your app and now you can take and compare
  screenshots.
* use jake upload-ssh-key to avoids crazy number of password prompts.
* Must restart the device the first installation of the agents.